Clean up download ahead logic
- Remove redundant chapter sorting logic when fetching next chapter(s)
- Remove redundant download queue checks (it'll handle already queued or downloaded items)
- Trigger download ahead when read >= 25% of chapter rather than 20%
- Rely on download cache when checking if next chapter is downloaded to avoid jank (fixes #8328)

(cherry picked from commit fc184f1cfa929328c41b0ec374503afe204d6c15)

# Conflicts:
#	app/src/main/java/eu/kanade/tachiyomi/data/library/LibraryUpdateService.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/reader/ReaderPresenter.kt

+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/advanced/design/MigrationBottomSheetDialog.kt
@@ -16,7 +16,7 @@ import eu.kanade.tachiyomi.R
 import eu.kanade.tachiyomi.core.preference.Preference
 import eu.kanade.tachiyomi.databinding.MigrationBottomSheetBinding
 import eu.kanade.tachiyomi.ui.browse.migration.MigrationFlags
-import eu.kanade.tachiyomi.util.system.toInt
+import eu.kanade.tachiyomi.util.system.toLong
 import eu.kanade.tachiyomi.util.system.toast
 import eu.kanade.tachiyomi.widget.sheet.BaseBottomSheetDialog
 import uy.kohesive.injekt.injectLazy
@@ -115,7 +115,7 @@ class MigrationBottomSheetDialog(private val activity: Activity, private val lis
      * Binds a radio group with a boolean preference.
      */
     private fun RadioGroup.bindToPreference(pref: Preference<Boolean>) {
-        (getChildAt(pref.get().toInt()) as RadioButton).isChecked = true
+        (getChildAt(pref.get().toLong().toInt()) as RadioButton).isChecked = true
         setOnCheckedChangeListener { _, value ->
             val index = indexOfChild(findViewById(value))
             pref.set(index == 1)
